[exchange] refactor 1.0
