+++ src/modules/ethereum/tradeTransactions.js
@@ -5,9 +5,9 @@ import moment from "moment";
 export async function fetchTradesTx(account, fromBlock, toBlock) {
     try {
         const exchangeInstance = store.getState().exchange.contract.ethersInstance;
-        const NewOrder = exchangeInstance.interface.events.NewOrder();
-        const OrderFill = exchangeInstance.interface.events.OrderFill();
-        const CancelledOrder = exchangeInstance.interface.events.CancelledOrder();
+        const NewOrder = exchangeInstance.interface.events.NewOrder;
+        const OrderFill = exchangeInstance.interface.events.OrderFill;
+        const CancelledOrder = exchangeInstance.interface.events.CancelledOrder;
         const provider = store.getState().web3Connect.ethers.provider;
 
         const paddedAccount = ethers.utils.hexlify(ethers.utils.padZeros(account, 32));
@@ -38,21 +38,19 @@ export async function fetchTradesTx(account, fromBlock, toBlock) {
             })
         ]);
 
-        const formattedLogsNewOrder = await Promise.all(
-            logsNewOrder.map(eventLog => _formatTradeLog(NewOrder, account, eventLog, 'NewOrder'))
-        );
-        const formattedLogsCanceledOrder = await Promise.all(
-            logsCanceledOrder.map(eventLog => _formatTradeLog(CancelledOrder, account, eventLog, 'CancelledOrder'))
-        );
-        const formattedLogsOrderFillBuy = await Promise.all(
-            logsOrderFillBuy.map(eventLog => _formatTradeLog(OrderFill, account, eventLog, 'OrderFillBuy'))
-        );
-        const formattedLogsOrderFillSell = await Promise.all(
-            logsOrderFillSell.map(eventLog => _formatTradeLog(OrderFill, account, eventLog, 'OrderFillSell'))
-        );
-        const logs = [...formattedLogsNewOrder, ...formattedLogsCanceledOrder, ...formattedLogsOrderFillBuy, ...formattedLogsOrderFillSell];
+        const logs = await Promise.all([
+            ...logsNewOrder.map(eventLog => _formatTradeLog(NewOrder, account, eventLog)),
+            ...logsCanceledOrder.map(eventLog => _formatTradeLog(CancelledOrder, account, eventLog)),
+            ...logsOrderFillBuy.map(eventLog => _formatTradeLog(OrderFill, account, eventLog)),
+            ...logsOrderFillSell.map(eventLog => _formatTradeLog(OrderFill, account, eventLog))
+        ]);
+
+        // const logs = [...formattedLogsNewOrder, ...formattedLogsCanceledOrder, ...formattedLogsOrderFillBuy, ...formattedLogsOrderFillSell];
+
+        console.log(logs);
 
         logs.sort((log1, log2) => {
+          console.log('log1',log1, 'log2', log2);
           return log2.blockData.timestamp - log1.blockData.timestamp;
         });
 
@@ -62,7 +60,7 @@ export async function fetchTradesTx(account, fromBlock, toBlock) {
     }
 }
 
-async function _formatTradeLog(event, account, eventLog, type) {
+async function _formatTradeLog(event, account, eventLog) {
     const decimalsDiv = store.getState().augmintToken.info.decimalsDiv;
     let blockData;
     if (typeof eventLog.getBlock === "function") {
@@ -84,12 +82,9 @@ async function _formatTradeLog(event, account, eventLog, type) {
     const ethAmountRounded = parseFloat(bn_ethAmount.toFixed(6));
     const tokenAmount = parseFloat(bn_tokenAmount / decimalsDiv);
     const price = parseFloat(parsedData.price / decimalsDiv);
-    const direction = tokenAmount === 0 ? -1 : 1;
 
     const logData = Object.assign({ args: parsedData }, eventLog, {
-        blockData: blockData,
-        direction: direction,
-        directionText: direction === -1 ? "sent" : "received",
+        blockData,
         blockTimeStampText: blockTimeStampText,
         bn_weiAmount: bn_weiAmount,
         bn_tokenAmount: bn_tokenAmount,
@@ -97,7 +92,7 @@ async function _formatTradeLog(event, account, eventLog, type) {
         ethAmount: ethAmount,
         ethAmountRounded: ethAmountRounded,
         price: price ,
-        type
+        type: event.name
     });
 
     return logData;
