+++ src/modules/reducers/exchange.js
@@ -1,5 +1,4 @@
 import store from "modules/store";
-import { fetchTradesTx } from "modules/ethereum/tradeTransactions";
 import SolidityContract from "modules/ethereum/SolidityContract";
 import exchangeArtifacts from "contractsBuild/Exchange.json";
 
@@ -11,12 +10,7 @@ export const EXCHANGE_REFRESH_REQUESTED = "exchange/EXCHANGE_REFRESH_REQUESTED";
 export const EXCHANGE_REFRESH_ERROR = "exchange/EXCHANGE_REFRESH_ERROR";
 export const EXCHANGE_REFRESH_SUCCESS = "exchange/EXCHANGE_REFRESH_SUCCESS";
 
-export const TRADE_REFRESH_REQUESTED = "trade/TRADE_REFRESH_REQUESTED";
-export const TRADE_REFRESH_ERROR = "trade/TRADE_REFRESH_ERROR";
-export const TRADE_REFRESH_SUCCESS = "trade/TRADE_REFRESH_SUCCESS";
-
 const initialState = {
-    contract: null,
     error: null,
     connectionError: false,
     isLoading: false,
@@ -80,32 +74,6 @@ export default (state = initialState, action) => {
                 info: action.info
             };
 
-        case TRADE_REFRESH_REQUESTED:
-            return {
-                isLoading: true,
-                account: action.account,
-                fromBlock: action.fromBlock,
-                toBlock: action.toBlock,
-                ...state
-            };
-
-        case TRADE_REFRESH_ERROR:
-            return {
-                ...state,
-                isLoading: false,
-                error: action.error
-            };
-
-        case TRADE_REFRESH_SUCCESS:
-            return {
-                ...state,
-                trades: action.result,
-                isLoading: false,
-                isConnected: true,
-                connectionError: false,
-                error: null
-            };
-
         default:
             return state;
     }
@@ -154,33 +122,6 @@ export const refreshExchange = () => {
         } catch (error) {
             return dispatch({
                 type: EXCHANGE_REFRESH_ERROR,
-                error: error
-            });
-        }
-    };
-};
-
-export const refreshTrades = (account, fromBlock, toBlock) => {
-    return async dispatch => {
-        dispatch({
-            type: TRADE_REFRESH_REQUESTED,
-            account: account,
-            fromBlock: fromBlock,
-            toBlock: toBlock
-        });
-        try {
-            const trades = await fetchTradesTx(account, fromBlock, toBlock);
-            const exchange = store.getState().exchange.contract.instance;
-            const info = await getExchangeInfo(exchange);
-
-            return dispatch({
-                type: TRADE_REFRESH_SUCCESS,
-                result: trades,
-                info: info
-            });
-        } catch (error) {
-            return dispatch({
-                type: TRADE_REFRESH_ERROR,
                 error: error
             });
         }
